requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
arabic-reshaper==3.0.0
async-timeout==5.0.1
attrs==25.3.0
auto_gptq==0.7.1
bcrypt==3.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
Brlapi==0.8.3
cachetools==6.1.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==44.0.2
cupshelpers==1.0
datasets==3.5.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.8
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
faiss-cpu==1.11.0
faiss-gpu==1.7.2
fastapi==0.115.12
fasteners==0.14.1
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
future==0.18.2
gekko==1.3.0
gitdb==4.0.12
GitPython==3.1.44
gpt4all==2.8.2
gradio==5.23.3
gradio_client==1.8.0
groovy==0.1.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama_cpp_python==0.3.8
llvmlite==0.44.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
narwhals==1.44.0
netifaces==0.11.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai-whisper==20240930
orjson==3.10.16
packaging==25.0
pandas==2.2.3
paramiko==2.9.3
pdfminer.six==20250327
peft==0.15.1
pexpect==4.8.0
Pillow==9.0.1
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==19.0.1
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyRFC3339==1.1
PySocks @ file:///home/hadise/Downloads/PySocks-1.7.1.tar.gz
python-apt==2.4.0+ubuntu4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
rich==14.0.0
rouge==1.0.1
rpds-py==0.25.1
ruff==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.10.0
sentence-transformers==4.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.1
streamlit==1.46.0
sympy==1.13.1
systemd-python==234
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tk==0.1.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.51.0
triton==3.2.0
ttystatus==0.38
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
wadllib==1.3.6
watchdog @ file:///home/hadise/Downloads/watchdog-3.0.0-py3-none-manylinux2014_x86_64.whl
websockets==15.0.1
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.19.0
zipp==1.0.0
langdetect